K_reread created (simplification much better)
Now i have allowed for rereads to happen. That is what was stopping the simplification from actually getting simple.

There are still some ways to make it simpler, which i'll be seeing next.

And, the code still needs a lot of cleaning now that K(), xK(), K_creator and other parts are not necessary

But now at least it's giving me much smaller expressions